Sample Data Cases

+++ Acme-Starters-C/src/main/java/acme/entities/strategy/Strategy.java
@@ -84,7 +84,7 @@ public class Strategy extends AbstractEntity {
 
 	@Transient
 	public Double getExpectedPercentaje() {
-		Integer result = this.repository.getExpectedPercentaje();
+		Double result = this.repository.getExpectedPercentaje();
 		return result;
 	}
