+++ Acme-Starters-C/src/main/java/acme/entities/strategy/StrategyRepository.java
@@ -9,5 +9,5 @@ import org.springframework.stereotype.Repository;
 public interface StrategyRepository extends CrudRepository<Strategy, Integer> {
 
 	@Query("SELECT SUM(t.expectedPorcentaje) FROM Tactic t")
-	Integer getExpectedPercentaje();
+	Double getExpectedPercentaje();
 }
